Moved ProxyServerTest into net.lightbody.bmp.proxy.test.util package
